# swagger.yaml
basePath: /api
definitions:
...
  server.CreateOrderRequest:
    properties:
      address:
        type: string
      amount:
        type: number
      email:
        type: string
      in_currency:
        type: string
      out_currency:
        type: string
      payment_address:
        type: string
    type: object
  server.CreateOrderResponse:
    properties:
      in_amount:
        type: number
      order_number:
        type: integer
      out_amount:
        type: number
      transfer_address:
        type: string
    type: object
  server.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  server.CurrentRateResponse:
    properties:
      amount:
        type: number
    type: object
...
  server.OrderStatusResponse:
    properties:
      status:
        type: string
    type: object
...
  server.UserOrder:
    properties:
      id:
        type: integer
      in_amount:
        type: number
      in_currency:
        type: string
      out_addr:
        type: string
      out_amount:
        type: number
      out_currency:
        type: string
      recv_addr:
        type: string
      status:
        type: string
    type: object
  server.UserOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/server.UserOrder'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Swap Support
    url: http://github.com/d1nch8g/swap
  description: Swap exchange api.
  license:
    name: MIT
    url: https://github.com/d1nch8g/swap/src/branch/main/LICENSE
  termsOfService: http://github.com/d1nch8g/swap
  title: Swap exchange api
  version: "1.0"
paths:
...
  /confirm-payment:
    post:
      parameters:
      - description: Order id
        in: query
        name: order_id
        required: true
        type: string
      - description: File with operation check
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
      summary: Confirm that payment operation is approved and provide check
  /create-order:
    post:
      parameters:
      - description: Request parameters
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateOrderRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CreateOrderResponse'
      summary: Create order to exchange specific currency
  /create-user:
    post:
      parameters:
      - description: Create user request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/server.CreateUserRequest'
      responses:
        "200":
          description: OK
      summary: Create new user request
  /current-rate:
    get:
      parameters:
      - description: Currency in
        in: path
        name: currency_in
        required: true
        type: string
      - description: Currency out
        in: path
        name: currency_out
        required: true
        type: string
      - description: Amount in
        in: path
        name: amount
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CurrentRateResponse'
      summary: Current rate at specific currency
...
  /login:
    post:
      parameters:
      - description: Email
        in: header
        name: email
        required: true
        type: string
      - description: Password
        in: header
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Login and get auth key
...
  /order-status:
    get:
      parameters:
      - description: orderid
        in: header
        name: Orderid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.OrderStatusResponse'
      summary: Check order status and return info about order
  /user/list-orders:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.UserOrdersResponse'
      security:
      - ApiKeyAuth: []
      summary: List user's orders
  /validate-card:
    post:
      parameters:
      - description: Email
        in: query
        name: email
        required: true
        type: string
      - description: Currency
        in: query
        name: currency
        required: true
        type: string
      - description: Address
        in: query
        name: addr
        required: true
        type: string
      - description: Approve file
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
      summary: Create order to exchange specific currency
  /verify/{uuid}:
    get:
      parameters:
      - description: UUID sent by email
        in: path
        name: uuid
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Verify user email address
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Token authorization for internal operations
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"